Backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
pydantic
whisper-openai
yt-dlp
ffmpeg-python
python-multipart
aiofiles
groq
pydantic-settings
supabase
python-jose[cryptography]
httpx